Creating ContactForm component
